EncounterCriteria: sbyte IVs
Narrow fields to reduce allocation < 16 bytes (maybe make this a struct?)
Add method to get combined IVs (gen3/4 format) for RNG purposes
use ^ method in RNG methods
Change GetExpectedLevel to be generic rather than interface, better codegen
Change Gender to enum for clarity

breaks ALM and potentially other plugins, just need to recompile if using IVs

+++ PKHeX.Core/Editing/IBattleTemplate.cs
@@ -53,7 +53,7 @@ public interface IBattleTemplate : ISpeciesForm, IGigantamaxReadOnly, IDynamaxLe
     /// <summary>
     /// <see cref="PKM.HPType"/> of the Set entity.
     /// </summary>
-    int HiddenPowerType { get; }
+    sbyte HiddenPowerType { get; }
 
     /// <summary>
     /// <see cref="EffortValues"/> of the Set entity.
